clean astronomical_objects_list code
