+++ multisurveys-apis/src/object_api/static/astronomical_object_list/conditions.js
@@ -1,0 +1,3 @@
+export function check_if_no_item_is_selected(object) {
+    return !object
+}
