+++ multisurveys-apis/src/object_api/templates/astronomical_object_list/astronomical_objects_list.html.jinja
@@ -9,15 +9,15 @@
     <input id="selected_oid_cache" class="tw-hidden" value="{{selected_oid}}">
 
 
-    <div class="tw-w-full tw-h-[95vh] tw-text-[0.875rem] dark:tw-bg-[#424242]">
-        <div class="tw-px-4 tw-text-black/60 tw-border-y tw-border-[#616161] dark:tw-text-white/70 tw-h-[45px] tw-flex tw-items-center tw-justify-start">Object List</div>
+    <div class="tw-w-full tw-h-[95vh] tw-text-[0.875rem] tw-flex tw-flex-col dark:tw-bg-[#424242]">
+        <div class="tw-px-4 tw-text-black/60 tw-border-y tw-border-[#616161] dark:tw-text-white/70 tw-grow tw-flex tw-items-center tw-justify-start">Object List</div>
 
         {% for item in objects_list['items']%}
             {% if item['oid']|string != selected_oid|string %}
                 <div
                 data-oid='{{item["oid"]}}'
                 name="sidebar-row-element" 
-                class="tw-h-[45px] tw-flex tw-justify-center tw-items-center obj-unselected-item"
+                class="tw-grow tw-flex tw-justify-center tw-items-center obj-unselected-item"
                 hx-on:click="highlight_new_object(event)"
                 >
                 {{item['oid']}}
@@ -27,7 +27,7 @@
                 <div
                  name="sidebar-row-element"
                  data-oid='{{item["oid"]}}'
-                 class="tw-flex tw-justify-center tw-items-center tw-h-[45px] obj-selected-item"
+                 class="tw-flex tw-justify-center tw-items-center tw-grow obj-selected-item"
                  hx-on:click="highlight_new_object(event)"
                  >
                     {{selected_oid}}
@@ -37,7 +37,7 @@
     </div>
 
 
-    <div class=" tw-flex tw-justify-center tw-items-center tw-text-base tw-h-[48px] dark:tw-bg-[#424242]">
+    <div class=" tw-flex tw-justify-center tw-items-center tw-text-base tw-h-[5vh] dark:tw-bg-[#424242]">
         {% if has_prev %}
             <span 
             id="prev_page"
@@ -85,7 +85,7 @@
 {% endif %}
 
 <script type="module">
-    import {init, elementReady} from "{{API_URL}}/static/sidebar.js"
+    import {init, elementReady} from "{{API_URL}}/static/astronomical_object_list/object_list.js"
 
 
     elementReady("#sidebar-objects-htmx").then((el) => {
